feat: added unmarshal and validate capabilities
added an auxiliar function to allow for linting directly out of a yaml
file content. this function parses the body into an intermediary struct
and then applies regular linting rules on top of the resulting object.

this commit also adds a considerable amount of changes in the linting
process, these changes were introduced after using the lint on top of a
database of 99k installer objects and capturing some problems with the
linting logic.